Split the `mixed` module
